requirements.txt
Flask==2.1.0
gevent==21.12.0
h5py==3.1.0
numpy==1.22.3
Pillow==8.3.2
tensorflow-cpu==2.8.0
Werkzeug==2.0.1